Working on upgradring the code to work with 2021
Had to change a few things from int to double to get it to complie.

Installed navx and pheonix on my end

+++ 2021AtHome/src/main/java/frc/robot/subsystems/Door.java
@@ -40,7 +40,7 @@ public class Door extends SubsystemBase {
     door.setSelectedSensorPosition(0);
 
   }
-  public int DoorPos(){
+  public double DoorPos(){
     return door.getSelectedSensorPosition();
   }
